dry a bit
